+++ app/src/main/java/com/livechatinc/livechatwidgetexample/FullScreenWindowActivityExample.java
@@ -47,8 +47,8 @@ public class FullScreenWindowActivityExample extends AppCompatActivity implement
         chatBadgeTv = findViewById(R.id.chat_badge);
         clearSessionBtn = findViewById(R.id.clear_session_btn);
         clearSessionBtn.setOnClickListener(view -> {
-            ChatWindowUtils.clearSession(view.getContext());
-            chatWindow.reload(false);
+            ChatWindowUtils.clearSession();
+            chatWindow.reload();
         });
     }