+++ inappchat/src/main/java/com/livechatinc/inappchat/FileChooserMode.java
@@ -1,11 +1,6 @@
 package com.livechatinc.inappchat;
 
-import android.os.Build;
-
-import androidx.annotation.RequiresApi;
-
-enum FileChooserMode {
+public enum FileChooserMode {
     SINGLE,
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
     MULTIPLE
 }